OBPIH-6213 Moved logic from StockMovementDisplayStatusUtil to StockMovementDisplayStatusHelper
