+++ grails-app/views/returns/show.gsp
@@ -130,7 +130,7 @@
                                 <g:message code="stockMovement.status.label"/>
                             </td>
                             <td class="value">
-                                <format:metadata obj="${stockMovement?.displayStatus ?: (stockMovement?.shipment?.status?.code ?: stockMovement?.statusCode)}"/>
+                                ${stockMovement?.displayStatus?.label}
                             </td>
                         </tr>
                         <tr class="prop">
