+++ grails-app/views/stockMovement/show.gsp
@@ -234,7 +234,7 @@
                                 <warehouse:message code="stockMovement.status.label"/>
                             </td>
                             <td class="value">
-                                <format:metadata obj="${stockMovement?.displayStatus}"/>
+                                ${stockMovement?.displayStatus?.label}
                             </td>
                         </tr>
                         <tr class="prop">
